LayoutTests/http/tests/security/cross-frame-access-parent-explicit-domain-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/resources/cross-frame-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-parent-explicit-domain.html. Domains, protocols and ports must match.


PASS: Cross frame access from frame explicitly setting document.domain was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.
